Editor Operation Manager, Block Add/Remove, and Refined Width/Depth Setting
Ported the editor operation manager from the other engine, which allows
the ability to undo/redo operations done in the editor. With this system
the adding and removing of blocks has been added. And some additional
functionality was added to width/depth setting; when setting these
values, they are now bounded by min/max values. Also, when resizing the
stage to be smaller, blocks out of range get removed.