Basic Editor UI
The basic editor UI has been laid out in the scene. Assets for building
that UI have been added (buttons and panel images). Pressing a button
does not cause a mouse raycast in the 3D scene. An example of opening up
menus in the editor has been added wtih the addition of the stage
settings sub menu.