Texture Experiments and Stage Offset
Includes experiments with how stage textures should be handled in Unity
(since there are so many and naive texture atlasing won't work due to
interpolation + mip-mapping). The stage settings now allow the user to
offset the stage blocks (to enable this, the stage is now held in a
1-dim array, which will be good for serializing the stage, and there now
exists a stage buffer array which can be used for various purposes).